split CBlockIndex in mutable and immutable data
CBlockIndex contains data that can never change (block header,
nHeight etc), it also contains data that can be updated, like
nStatus.

If one updated some data, one had to write all data back to the
database.

This patch stores immutable and mutable data under different keys
in leveldb, so that not all immutable data needs to be in memory
if the mutable data needs to be updated.