Add DNSSeed i0seed.snel.it
i0seed.snel.it runs bind with phpdnsseed.php from contrib/phpdnsseed/
It is a simple system that extract connectable nodes from a running
i0coind en converts the list in a bind zonefile. This zonefile is moved
to /etc/bind and a reload of bind is triggered. (every 10 minutes)