fix retarget at 890000
I0coins old retarget algorithm is asymmetric; difficulty decrease
is fast and difficulty increase is slow. This can be abused to
make a 51% attack more profitable than it should be (thereby
encouraging it).

This patch changes the retarget algorithm to be on par with
bitcoin's alogirithm (max factor of 4 every 2016 blocks).
Since i0coin retargets every 120 blocks, the actual retarget
factor is 4^(120/2016) = 1.086.

Use integer arithmetic, to make sure that all architectures
return the exact same answer.